requirements.txt
certifi==2018.11.29
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
mysql-connector==2.2.9
